Add a configuration option to show idle timer.
When an activity completed, and the user didn't pay attention to the
notifications, it is impossible for the user to visualize how much time
has elapsed since the activity completed. Hence, adding the idle
timer.

This will also be useful for users who gets distracted too easily, as
there will be visual reminder of how long they have been distracted
for without following the flow.