Add extensions support.
Wanting to integrate with TaskWarrior motivated me to add extensions support for
Zendoro to keep the core small, and still make it extensible. The initial
extension system is very primitive, and currently TaskWarrior and Stats are
added as default extensions. Later, there will be some cleanup made, and these
things might be extracted to a separate repo.